Change how attribute mask hack works
Because of changes upstream, the way it had to be done was changed.

But TBH I don't think it worked properly in the first place,
because it seems that VertexAttribBufferMask isn't always
a subset of _Enabled.

Now we instead disable all user space arrays in _mesa_update_vao_derived_arrays.